requirements.txt
fastapi==0.111.1
python-dotenv==1.0.1
qrcode==7.4.2